+++ lib/pocketbase.dart
@@ -13,6 +13,7 @@ export "src/dtos/admin_model.dart";
 export "src/dtos/apple_client_secret.dart";
 export "src/dtos/auth_method_provider.dart";
 export "src/dtos/auth_methods_list.dart";
+export "src/dtos/backup_file_info.dart";
 export "src/dtos/collection_model.dart";
 export "src/dtos/external_auth_model.dart";
 export "src/dtos/health_check.dart";
@@ -27,6 +28,7 @@ export "src/dtos/schema_field.dart";
 
 // services (exported mainly for dartdoc - https://github.com/dart-lang/dartdoc/issues/2154)
 export "src/services/admin_service.dart";
+export "src/services/backup_service.dart";
 export "src/services/base_crud_service.dart";
 export "src/services/collection_service.dart";
 export "src/services/file_service.dart";