+++ lib/src/dtos/health_check.g.dart
@@ -9,10 +9,12 @@ part of 'health_check.dart';
 HealthCheck _$HealthCheckFromJson(Map<String, dynamic> json) => HealthCheck(
       code: json['code'] as int? ?? 0,
       message: json['message'] as String? ?? "",
+      data: json['data'] as Map<String, dynamic>? ?? const {},
     );
 
 Map<String, dynamic> _$HealthCheckToJson(HealthCheck instance) =>
     <String, dynamic>{
       'code': instance.code,
       'message': instance.message,
+      'data': instance.data,
     };